msuss/requirements.txt
google-generativeai
pytest
